Update GetReportQuery Handler
